\name{KRMM-package}
\alias{KRMM-package}
\docType{package}
\encoding{utf8}

\title{
  Kernel Ridge Mixed Model (KRMM) Package
}

\description{
  The \code{KRMM} package provides tools for solving kernel ridge regression within the mixed model framework. It supports various kernels including linear, polynomial, Gaussian, Laplacian, and ANOVA. The package uses the expectation-maximization (EM) algorithm to estimate model components (fixed and random effects) and variance parameters. Estimated components and parameters such as BLUP of dual variables and BLUP of random predictor effects for the linear kernel (also known as RR-BLUP) are available.
}

\details{
  The \code{KRMM} package implements kernel ridge regression for various kernels in the mixed model framework defined by \eqn{Y = X\beta + Zu + \epsilon}, where \eqn{X} and \eqn{Z} are the design matrices of predictors with fixed and random effects, respectively. The package provides the following functions: \code{krmm}, \code{predict_krmm}, \code{tune_krmm}, and \code{em_reml_mm}.
}

\author{
  Laval Jacquin

  Maintainer: Laval Jacquin <jacquin.julien@gmail.com>
}

\references{
  Jacquin et al. (2016). A Unified and Comprehensible View of Parametric and Kernel Methods for Genomic Prediction with Application to Rice. Front. Genet. 7:145 (in peer review)

  Robinson, G. K. (1991). That blup is a good thing: the estimation of random effects. Statistical Science, 534 15-32

  Foulley, J.-L. (2002). Algorithme em: théorie et application au modèle mixte. Journal de la Société française de Statistique 143, 57-109
}

\keyword{ package }


\examples{

# load libraries
library(KRMM)
library(MASS)
library(kernlab)
library(cvTools)

# simulate data
set.seed(123)
p <- 100
n <- 200
beta <- rnorm(p, mean = 0, sd = 1.0) # random effects
X <- matrix(runif(p * n, min = 0, max = 1), ncol = p, byrow = T) # matrix of covariates
f <- tcrossprod(beta, X) # data generating process
eps <- rnorm(n, mean = 0, sd = 0.9) # add residuals
Y <- f + eps

# split data into training and test set
n_train <- floor(n * 0.67)
idx_train <- sample(1:n, size = n_train, replace = F)

# train
x_train <- X[idx_train, ]
y_train <- Y[idx_train]
length(y_train)

# test
x_test <- X[-idx_train, ]
y_test <- Y[-idx_train]
f_test <- f[-idx_train] # true value generated by DGP we want to predict
length(y_test)

# train and predict with krmm linear kernel
linear_krmm_model <- krmm(
  Y = y_train, Matrix_covariates = x_train,
  method = "RR-BLUP"
)

# without fixed effects
f_hat_test <- predict_krmm(linear_krmm_model,
  Matrix_covariates = x_test
)

dev.new()
plot(f_hat_test, f_test,
  main = "Linear RKHS regression with default rate of decay (not optimized)"
)
cor(f_hat_test, f_test)

# add fixed effects
f_hat_test <- predict_krmm(linear_krmm_model,
  Matrix_covariates = x_test, add_flxed_effects = T
)

dev.new()
plot(f_hat_test, f_test,
  main = "Linear RKHS regression with default rate of decay (not optimized)"
)
cor(f_hat_test, f_test)

# train and predict with krmm gaussian kernel (default kernel for RKHS method)
non_linear_krmm_model <- krmm(
  Y = y_train, Matrix_covariates = x_train,
  method = "RKHS"
)

# without fixed effects
f_hat_test <- predict_krmm(non_linear_krmm_model,
  Matrix_covariates = x_test
)

dev.new()
plot(f_hat_test, f_test,
  main = "Gaussian RKHS regression with default rate of decay (not optimized)"
)
cor(f_hat_test, f_test)

# add fixed effects
f_hat_test <- predict_krmm(non_linear_krmm_model,
  Matrix_covariates = x_test, add_flxed_effects = T
)

dev.new()
plot(f_hat_test, f_test,
  main = "Gaussian RKHS regression with default rate of decay (not optimized)"
)
cor(f_hat_test, f_test)

\dontrun{

# tune krmm model with a gaussian kernel
non_linear_opt_krmm_obj <- tune_krmm(
  Y = y_train, Matrix_covariates = x_train,
  rate_decay_grid = seq(0.01, 0.1, length.out = 5), nb_folds = 3,
  method = "RKHS"
)
non_linear_opt_krmm_obj$optimal_h

plot(non_linear_opt_krmm_obj$rate_decay_grid,
  non_linear_opt_krmm_obj$mean_loss_grid,
  type = "l"
)

non_linear_opt_krmm_model <- non_linear_opt_krmm_obj$optimized_model

# add fixed effects
f_hat_test <- predict_krmm(non_linear_opt_krmm_model,
  Matrix_covariates = x_test, add_flxed_effects = T
)
dev.new()
plot(f_hat_test, f_test,
  main = "Gaussian RKHS regression with optimized rate of decay"
)
cor(f_hat_test, f_test)

}

}
